Moved CurrentValueState to its own file.
